updates BFS & DFS
